+++ cortex-js/src/infrastructure/commanders/cortex-command.commander.ts
@@ -11,11 +11,11 @@ import pkg from '@/../package.json';
 import { PresetCommand } from './presets.command';
 import { TelemetryCommand } from './telemetry.command';
 import { SetCommandContext } from './decorators/CommandContext';
-import { ContextService } from '@/util/context.service';
 import { EmbeddingCommand } from './embeddings.command';
 import { BenchmarkCommand } from './benchmark.command';
 import chalk from 'chalk';
 import { printSlogan } from '@/utils/logo';
+import { ContextService } from '../services/context/context.service';
 
 @RootCommand({
   subCommands: [
