+++ cortex-js/src/infrastructure/commanders/models/model-pull.command.ts
@@ -1,9 +1,9 @@
 import { CommandRunner, SubCommand } from 'nest-commander';
 import { exit } from 'node:process';
 import { SetCommandContext } from '../decorators/CommandContext';
-import { ContextService } from '@/util/context.service';
 import { ModelsCliUsecases } from '@commanders/usecases/models.cli.usecases';
 import { ModelNotFoundException } from '@/infrastructure/exception/model-not-found.exception';
+import { ContextService } from '@/infrastructure/services/context/context.service';
 
 @SubCommand({
   name: 'pull',
