+++ cortex-js/src/infrastructure/commanders/models/model-remove.command.ts
@@ -1,8 +1,8 @@
 import { CommandRunner, SubCommand } from 'nest-commander';
 import { ModelsCliUsecases } from '@commanders/usecases/models.cli.usecases';
 import { exit } from 'node:process';
-import { ContextService } from '@/util/context.service';
 import { SetCommandContext } from '../decorators/CommandContext';
+import { ContextService } from '@/infrastructure/services/context/context.service';
 
 @SubCommand({
   name: 'remove',
