+++ cortex-js/src/infrastructure/commanders/presets.command.ts
@@ -3,7 +3,7 @@ import { readdirSync } from 'fs';
 import { CommandRunner, SubCommand } from 'nest-commander';
 import { join } from 'path';
 import { SetCommandContext } from './decorators/CommandContext';
-import { ContextService } from '@/util/context.service';
+import { ContextService } from '../services/context/context.service';
 
 @SubCommand({
   name: 'presets',
