+++ cortex-js/src/infrastructure/middlewares/app.logger.middleware.ts
@@ -1,8 +1,8 @@
 import { TelemetrySource } from '@/domain/telemetry/telemetry.interface';
-import { ContextService } from '@/util/context.service';
 import { Injectable, NestMiddleware, Logger } from '@nestjs/common';
 
 import { Request, Response, NextFunction } from 'express';
+import { ContextService } from '../services/context/context.service';
 
 @Injectable()
 export class AppLoggerMiddleware implements NestMiddleware {
