+++ cortex-js/src/usecases/models/models.module.ts
@@ -5,10 +5,10 @@ import { CortexModule } from '@/usecases/cortex/cortex.module';
 import { ExtensionModule } from '@/infrastructure/repositories/extensions/extension.module';
 import { HttpModule } from '@nestjs/axios';
 import { TelemetryModule } from '../telemetry/telemetry.module';
-import { UtilModule } from '@/util/util.module';
 import { FileManagerModule } from '@/infrastructure/services/file-manager/file-manager.module';
 import { ModelRepositoryModule } from '@/infrastructure/repositories/models/model.module';
 import { DownloadManagerModule } from '@/infrastructure/services/download-manager/download-manager.module';
+import { ContextModule } from '@/infrastructure/services/context/context.module';
 
 @Module({
   imports: [
@@ -18,7 +18,7 @@ import { DownloadManagerModule } from '@/infrastructure/services/download-manage
     HttpModule,
     FileManagerModule,
     TelemetryModule,
-    UtilModule,
+    ContextModule,
     ModelRepositoryModule,
     DownloadManagerModule,
   ],
