+++ cortex-js/src/usecases/telemetry/telemetry.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { TelemetryUsecases } from './telemetry.usecases';
 import { HttpModule } from '@nestjs/axios';
 import { DatabaseModule } from '@/infrastructure/database/database.module';
-import { UtilModule } from '@/util/util.module';
 import { FileManagerModule } from '@/infrastructure/services/file-manager/file-manager.module';
+import { ContextModule } from '@/infrastructure/services/context/context.module';
 
 @Module({
-  imports: [HttpModule, DatabaseModule, FileManagerModule, UtilModule],
+  imports: [HttpModule, DatabaseModule, FileManagerModule, ContextModule],
   providers: [TelemetryUsecases],
   exports: [TelemetryUsecases],
 })
